+++ CHANGELOG.md
@@ -8,6 +8,15 @@ Version changes are pinned to SDK releases.
 test
 
 - error: Move error related functionality to error.ts. ([#6](https://github.com/zetamarkets/sdk/pull/6))
+- insurance-client: Insurance fund functionality, whitelist checks, deposit & withdrawal. ([#9](https://github.com/zetamarkets/sdk/pull/9))
+- exchange: Insurance functionality, whitelist a user, rebalance vaults. ([#9](https://github.com/zetamarkets/sdk/pull/9))
+
+## [0.7.3] 2021-11-29
+
+- error: Move error related functionality to error.ts. ([#6](https://github.com/zetamarkets/sdk/pull/6))
+- client: Add `CancelMultipleOrders`. ([#8](https://github.com/zetamarkets/sdk/pull/8))
+
+Note: Skipped versions due to NPM issues.
 
 ## [0.7.0] 2021-11-17